+++ v3/crates/engine/src/routes/jsonapi.rs
@@ -143,6 +143,10 @@ async fn handle_rest_request(
                                                                          // we tell the user, for
                                                                          // now default to nothing
             ),
+            jsonapi::RequestError::PlanError(plan::PlanError::Relationship(msg)) => (
+                axum::http::StatusCode::BAD_REQUEST,
+                Json(serde_json::json!({"error": msg })),
+            ),
             jsonapi::RequestError::ExecuteError(field_error) => (
                 axum::http::StatusCode::INTERNAL_SERVER_ERROR,
                 Json(serde_json::json!({"error": field_error.to_string() })),