+++ v3/crates/jsonapi/src/catalog.rs
@@ -1,4 +1,4 @@
 mod types;
-pub use types::{Catalog, Model, ObjectType, State, Type};
+pub use types::{Catalog, Model, ObjectType, RelationshipTarget, State, Type};
 mod models;
 mod object_types;