+++ v3/crates/open-dds/src/query.rs
@@ -102,7 +102,7 @@ pub struct RelationshipSelection {
     #[serde(flatten)]
     pub target: RelationshipTarget,
     /// If the relationship output produces an object type or an array of object types, what to select from that/those object(s).
-    pub selection: Option<IndexMap<String, ObjectSubSelection>>,
+    pub selection: Option<IndexMap<Alias, ObjectSubSelection>>,
 }
 
 #[derive(Serialize, Deserialize, Debug, Clone, PartialEq, Eq)]