+++ v3/crates/plan/src/types.rs
@@ -2,5 +2,6 @@
 pub enum PlanError {
     Internal(String),   // equivalent to DataFusionError::Internal
     Permission(String), // equivalent to DataFusionError::Plan
+    Relationship(String),
     External(Box<dyn std::error::Error + Send + Sync>), //equivalent to DataFusionError::External
 }