+++ v3/crates/sql/src/execute/planner/common.rs
@@ -16,7 +16,7 @@ impl PhysicalPlanOptions {
 pub fn from_plan_error(plan: PlanError) -> DataFusionError {
     match plan {
         PlanError::Internal(msg) => DataFusionError::Internal(msg),
-        PlanError::Permission(msg) => DataFusionError::Plan(msg),
+        PlanError::Permission(msg) | PlanError::Relationship(msg) => DataFusionError::Plan(msg),
         PlanError::External(error) => DataFusionError::External(error),
     }
 }