+++ v3/changelog.md
@@ -4,6 +4,9 @@
 
 ### Added
 
+- Added support for fetching relationships in JSON:API using the `include`
+  parameter.
+
 ### Fixed
 
 - Fixed an error that occurred when filtering by more than one nested field at a